+++ local/shims/os.js
@@ -1,0 +1,3 @@
+export function hostname() { return 'this-device'; }
+export function networkInterfaces() { return {}; }
+export default { hostname, networkInterfaces };