+++ local/shims/package.js
@@ -1,0 +1,2 @@
+export const version = (typeof SUDS_VERSION !== 'undefined' ? SUDS_VERSION : 'local');
+export default { version };